+++ src/app/office/page.module.css
@@ -1,11 +1,10 @@
-body {
-  background-color: var(--clr-office-50);
-}
-
 .office__page {
+  background-color: var(--clr-office-50);
   display: flex;
   flex-direction: row;
   font-family: Helvetica, sans-serif;
+  width: 100vw;
+  height: 100vh;
 }
 
 .office__sidebar {
